+++ src/primary.js
@@ -16,7 +16,8 @@ inherit(FromBinder, Stream, {
       , emitter = {
         emit: function(x) {  $._send(VALUE, x, isCurrent)  },
         error: function(x) {  $._send(ERROR, x, isCurrent)  },
-        end: function() {  $._send(END, null, isCurrent)  }
+        end: function() {  $._send(END, null, isCurrent)  },
+        emitEvent: function(e) {  $._send(e.type, e.value, isCurrent)  }
       };
     this._unsubscribe = this._fn(emitter) || null;
 
@@ -70,6 +71,9 @@ inherit(Emitter, Stream, {
   end: function() {
     this._send(END);
     return this;
+  },
+  emitEvent: function(event) {
+    this._send(event.type, event.value);
   }
 });
 
@@ -130,4 +134,81 @@ inherit(ConstantError, Property, {
 Kefir.constantError = function(x) {
   return new ConstantError(x);
 }
+
+
+
+
+// Kefir.repeat(generator)
+
+function Repeat(generator) {
+  Stream.call(this);
+  this._generator = generator;
+  this._source = null;
+  this._inLoop = false;
+  this._activating = false;
+  this._iteration = 0;
+
+  var $ = this;
+  this._$handleAny = function(event) {
+    $._handleAny(event);
+  };
+}
+
+inherit(Repeat, Stream, {
+
+  _name: 'repeat',
+
+  _handleAny: function(event) {
+    if (event.type === END) {
+      this._source = null;
+      this._startLoop();
+    } else {
+      this._send(event.type, event.value, this._activating);
+    }
+  },
+
+  _startLoop: function() {
+    if (!this._inLoop) {
+      this._inLoop = true;
+      while (this._source === null && this._alive && this._active) {
+        this._source = this._generator(this._iteration++);
+        if (this._source) {
+          this._source.onAny(this._$handleAny);
+        } else {
+          this._send(END);
+        }
+      }
+      this._inLoop = false;
+    }
+  },
+
+  _onActivation: function() {
+    this._activating = true;
+    if (this._source) {
+      this._source.onAny(this._$handleAny);
+    } else {
+      this._startLoop();
+    }
+    this._activating = false;
+  },
+
+  _onDeactivation: function() {
+    if (this._source) {
+      this._source.offAny(this._$handleAny);
+    }
+  },
+
+  _clear: function() {
+    Stream.prototype._clear.call(this);
+    this._generator = null;
+    this._source = null;
+    this._$handleAny = null;
+  }
+
+});
+
+Kefir.repeat = function(generator) {
+  return new Repeat(generator);
+}
+